backend/requirements.txt
Flask
google-genai
python-dotenv
flake8
requests==2.31.0